+++ test/test_request.py
@@ -47,6 +47,12 @@ class TestRequest(unittest.TestCase):
             self.assertFalse(self.STATUS.Is_cancelled())
         except NotImplementedError:
             self.skipTest('mpi-status-is_cancelled')
+        flag = self.REQUEST.get_status()
+        self.assertTrue(flag)
+        flag = self.REQUEST.get_status(self.STATUS)
+        self.assertEqual(self.STATUS.source, MPI.ANY_SOURCE)
+        self.assertEqual(self.STATUS.tag,    MPI.ANY_TAG)
+        self.assertEqual(self.STATUS.error,  MPI.SUCCESS)
 
 
 class TestRequestArray(unittest.TestCase):